+++ flutter-nest-househelp-master/src/bookings/exceptions/invalid-booking-duration.exception.ts
@@ -1,0 +1,8 @@
+import { BadRequestException } from '@nestjs/common';
+
+export class InvalidBookingDurationException extends BadRequestException {
+  constructor(message: string) {
+    super(message);
+    this.name = 'InvalidBookingDurationException';
+  }
+}